Separate single Cypress command for mocking API endpoints into general, project and model mock commands.
