+++ cypress/e2e/about.cy.js
@@ -1,6 +1,6 @@
 describe('The About Page', () => {
   it('successfully loads', () => {
-    cy.mockApiRoutes();
+    cy.mockCommonApiEndpoints();
     cy.visit('/about');
     cy.get('body');
     cy.get('header');
