+++ cypress/e2e/home.cy.js
@@ -1,7 +1,7 @@
 // / <reference types="Cypress" />
 describe('The Home Page', () => {
   it('successfully loads', () => {
-    cy.mockApiRoutes();
+    cy.mockCommonApiEndpoints();
     cy.visit('/');
     cy.get('body');
   });
