+++ cypress/e2e/project/keyboard-shortcuts.cy.js
@@ -18,7 +18,7 @@ const instance = {
 
 describe('Test keyboard shortcuts', () => {
   beforeEach(() => {
-    cy.mockApiRoutes();
+    cy.mockCommonApiEndpoints();
     cy.fakeLogin();
     cy.setWebsocketWorkflow('websocket-workflow/retrain.json');
