+++ cypress/e2e/project/layers-panel.cy.js
@@ -1,6 +1,6 @@
 describe('Open existing project', () => {
   beforeEach(() => {
-    cy.mockApiRoutes();
+    cy.mockCommonApiEndpoints();
   });
 
   it('successfully loads', () => {
