+++ cypress/e2e/project/panel.cy.js
@@ -16,7 +16,7 @@ const instance = {
 
 describe('Panel functions', () => {
   beforeEach(() => {
-    cy.mockApiRoutes();
+    cy.mockCommonApiEndpoints();
     cy.fakeLogin();
     cy.setWebsocketWorkflow('websocket-workflow/retrain.json');
